Remove old Picard**Iterators (#1065)
* rename and remove

* part excision

* will break examples but need to move
