+++ examples/equilibria/eudemo_2017.py
@@ -47,7 +47,7 @@ from bluemira.equilibria.opt_constraints import AutoConstraints
 from bluemira.equilibria.optimiser import BreakdownOptimiser, FBIOptimiser
 from bluemira.equilibria.physics import calc_beta_p_approx, calc_li, calc_psib
 from bluemira.equilibria.profiles import BetaIpProfile, CustomProfile, DoublePowerFunc
-from bluemira.equilibria.solve import PicardAbsIterator, PicardLiAbsIterator
+from bluemira.equilibria.solve import PicardIterator
 
 # %%[markdown]
