+++ examples/equilibria/single_null_new.py
@@ -55,7 +55,7 @@ from bluemira.equilibria.opt_problems import (
     UnconstrainedTikhonovCurrentCOP,
 )
 from bluemira.equilibria.profiles import CustomProfile
-from bluemira.equilibria.solve_new import DudsonConvergence, PicardIterator
+from bluemira.equilibria.solve import DudsonConvergence, PicardIterator
 from bluemira.utilities.opt_problems import OptimisationConstraint
 from bluemira.utilities.optimiser import Optimiser
